v.0.7.3.3 Xamarin (no modifications to Win Forms version). Fixes to behaviour in Foods management and choose for Meal Page. Warning prompts on every "SAVE" button. Added notes management in glucose measurement page (field was already in the database).

+++ SharedGlucoMan/BusinessLayer/BL_General.cs
@@ -71,8 +71,13 @@ namespace SharedGlucoMan.BusinessLayer
         {
             try
             {
-                string backupFile = Path.Combine(Common.PathImportExport, Common.DatabaseFileName.Replace(".Sqlite", "_backup.Sqlite")); 
-                File.Copy(pathAndFileInternalDatabase, backupFile, true);
+                string backupFile = Path.Combine(Common.PathImportExport, Common.DatabaseFileName.Replace(".Sqlite", "_backup.Sqlite"));
+                if (File.Exists(pathAndFileInternalDatabase))
+                {
+                    File.Copy(pathAndFileInternalDatabase, backupFile, true);
+                    //File.Delete(pathAndFileInternalDatabase);
+                    //File.Create(pathAndFileInternalDatabase); 
+                }
                 File.Copy(pathAndFileExternalDatabase, pathAndFileInternalDatabase, true);
                 return true;
             }
